adds readme and fixes import task

+++ db/migrate/20181127200428_create_items.rb
@@ -3,7 +3,7 @@ class CreateItems < ActiveRecord::Migration[5.2]
     create_table :items do |t|
       t.string :name
       t.text :description
-      t.integer :unit_price
+      t.bigint :unit_price
       t.references :merchant, foreign_key: true
 
       t.timestamps
